requirements.txt
picamera2==0.3.17
openai==1.12.0
opencv-python==4.9.0.80
elevenlabs==0.2.27
resend==0.8.0
python-dotenv==0.21.0
torch==2.2.1
torchvision==0.17.1